+++ acvm/src/pwg/brillig.rs
@@ -127,8 +127,8 @@ impl BrilligSolver {
                     insert_value(witness, FieldElement::zero(), initial_witness)?
                 }
                 BrilligOutputs::Array(witness_arr) => {
-                    for w in witness_arr {
-                        insert_value(w, FieldElement::zero(), initial_witness)?
+                    for witness in witness_arr {
+                        insert_value(witness, FieldElement::zero(), initial_witness)?
                     }
                 }
             }